[feat/#3]:과목선택 시 저장가능
